Use our script to run the auth dance
